squash security patches
This is an illustration of an idea for making the history of branches that have security patches applied a little easier to follow, by squashing the commits contained in the patches into a single commit.

The idea here is that instead of dumping a collection of random commits into the branch history (with random dates) we can present a more meaningful history while still keeping the details of the specific commits that went into each patch.

We could take this approach a step further by squashing each patch individually instead of applying them all together, which would make it much easier to see from the branch history which security patches have been applied. That would just require moving the existing code into a loop over the patch files.